creates class module for calculating performance and uses it inline in the amountFor function
