+++ 01/PerformanceCalculator.ts
@@ -1,0 +1,33 @@
+import { Play, PlayPerformance } from './types.ts';
+
+export class PerformanceCalculator {
+  performance: PlayPerformance;
+  play: Play;
+
+  constructor(aPerformance: PlayPerformance, aPlay: Play) {
+    this.performance = aPerformance;
+    this.play = aPlay;
+  }
+
+  get amount() {
+    let result = 0;
+    switch (this.play.type) {
+      case 'tragedy':
+        result = 40000;
+        if (this.performance.audience > 30) {
+          result += 1000 * (this.performance.audience - 30);
+        }
+        break;
+      case 'comedy':
+        result = 30000;
+        if (this.performance.audience > 20) {
+          result += 10000 + 500 * (this.performance.audience - 20);
+        }
+        result += 300 * this.performance.audience;
+        break;
+      default:
+        throw new Error(`unknown type: ${this.play.type}`);
+    }
+    return result;
+  }
+}
